Add dashboard, indexes, and numeric type changes
Introduce dashboard API and service, add DB indexes and change several numeric types. Added DashboardController, IDashboardService and DashboardService, and registered DashboardService in Program. Updated DiabitsDbContext to create indexes on UserId/StartTime for GlucoseLevels, HeartRates, SleepSessions, Steps, Workouts, Menstruation and Medications. Added EF migrations (AddIndexes, ChangeDecimalValueTypes) to apply indexes and migrate decimal properties to double/float.

Changed numeric model types: GlucoseLevel.mmolL, Medication.Quantity and Medication.StrengthValue from decimal to double; updated model snapshot and migrations accordingly. Updated MapperFactory with a FromDto overload to map HealthDataPointBaseDto polymorphically. Made ManualInputDto.Flow nullable and added a commented TimelineResponse in DTOs.

Refactored auth refresh flow: removed userId parameter from IAuthService.RefreshAccessTokenAsync and updated AuthService and AuthController to use the token to look up the user; tests touching refresh token behavior were commented/adjusted. Also updated test expectations for mmolL to use double. Minor test and helper adjustments to match API changes.

+++ Diabits.API/Models/HealthDataPoints/HealthConnect/GlucoseLevel.cs
@@ -5,5 +5,5 @@
 /// </summary>
 public class GlucoseLevel : HealthDataPoint
 {
-    public decimal mmolL { get; set; }
+    public double mmolL { get; set; }
 }
